Initialize Simple Types "as in Java"
All class attributes are now initialized in the declaration with
curly braces which are empty in case of no default value literal.
The intention is to initialize even simple members "the same way
as in Java", which eases the processing of foreign ecore models and
instances, where the author has omitted the specification of a
default value and relied on the Java type system.

Additional changes:

- The ecore model code and the tests were recreated.

- The bootstrap script in XMLType now calls the code generator with
  an emf4cpp (library) path relative to the source code path which gets
  created when the library is build with the 'generator-release.sh'
  script.
  With this option the 'link_directories' in the type.cmake are set
  properly so that the 'emf4cpp-ecore' library will be found when linking
  the 'emf4cpp-type' library.

- The EMF4CPP path was set relative to the current code path in 'update-examples.sh'
  to omit policy warnings when creating the makefiles for the tests.

- The initialization of reference (smart) pointers was removed
  from the inititialization lists because they get initialized
  anyways.