Extend item provider getText() method with optional EObject_ptr
In the current implementation of the item providers the getText() methods
either take a EObject_ptr or a EClass_ptr, EAttribute_ptr or EStructuralFeature_ptr.
With the latter it is not possible to customize the output of _getText()
according to the context of the EObject instance in derived class
implementations.

To overcome this restriction the ItemProvider::getText() methods now
have an additional (optional) parameter for an EObject_ptr that can
be used to retrieve context information in derived implementations.