Replace boost::intrusive_ptr by std::shared_ptr
This commit replaces boost::intrusive_ptr by std::shared_ptr
in the library. All code regarding reference counting was
removed to and the includes were cleaned up.

EObject and Resource now derive from std::enable_shared_from_this().
This is needed for the implementation of the _this() methods
of EObject and its derivatives and to set the eResource of an
EObject added to the content list of the Resource.

The explicit removal of a Resource from the content list of a
ResourceSet was removed from the destructor of the Resource because
this is a containment relation and the destructor will not be
called until the Resource will be removed from the content
list anyways.