Replace plain structural feature c-pointer in EList
All occurences of EStructuralFeature* in the EList interface were
replaced by const references to ::ecore::EstructuralFeature_ptr and
the implementation of the FeatureEListImpl was adapted.

The CrossReferencer now also uses smart pointers to features.

Additional changes:

Regenerated all tests and libraries because the PackageFactory
template of the code generator was changed in a previous commit.