Correctly Dissolve Opposite Relation
Let LeftHand have a feature rightee with multiplicity one and let
RightHand have a feature leftees with multiplicity many.
LeftHand.rightee and RightHand.leftees are in an opposite relation. Let
lh be an instance of LeftHand, rh1 and rh2 be instances of RightHand,
lh.rightee point to rh1, and rh1.leftees contain lh. When lh is added to
rh2.leftees, then lh.rightee points to rh2, but rh1.leftees still
contains lh, which is incorrect. This removes lh from rh1.leftees.