Better FeatureId Assignment
Up to now emf4cpp was generating a unique id
for each structural feature defined in the ecore.
This works as long as no cross ecore inheritance
comes into play. If the base class lives in one
ecore and the derived class lives in another their
structural feature ids might collide in the
eSet/eGet methods of the derived class.

To overcome this restriction the feature id count
is now local to a classifier. It starts with the
base class structural features (in the order of
the super type list) followed by the structural
features of actual class in the order of their
definition in the ecore. This has the immediate
consequence that the same structural feature
might have different integer ids in different
classes.

The id of a feature of a class instance can now be
retrieved by calling the method
'EInt EClass::getFeatureID(EStructuralFeature_ptr)'
of the EClass instance. This function is also used
in the EObject::eGet()/eSet() methods and in the new
EObject::_inverseAdd()/_inverseRemove() methods.

Since the feature id is not unique across the
package anymore it can not be used as a constant
in a switch-case construct for structural features.
A static constexpr helper function:
'int64_t EPackage::makeUniqueFeatureID(int32_t classID,
int32_t featureID)'
was added to create a unique id from the id
of the feature and the id of the containing class.
This is used in the code generated for the item
providers.

The initialization of the ELists of all structural
features and the mapping from the structural
feature pointer to the feature id was moved from
the central method 'init_collections()' to the
method EClass::getEAllStructuralFeatures() at the
first call. This is necessary to ensure that all
(base class) packages are initialized properly.
The same holds true for all other getters
(references, attributes, operations etc).

As a consequence the constant reference getters
now use the non-const getters by casting away
the constness of the object if the reference list
is not initialized.

It is very important that the initialization of
the structural feature mapping only takes place
after the initialization of the packages. This
has also be ensured for references with opposite
relations because of the implicit calls to the
_inverseAdd() methods which also trigger calls
to EClass::getEAllStructuralFeatures() for classes
of not yet initialized packages. Therefore the
basicset()/Add() methods are now used in
XYPackage::_initPackage() and the opposite reference
is set 'by hand'.

A new test 'crossEcoreInheritance' was added to
the emf4cpp.tests folder. It consists of two
ecores where classes of the 'derived.ecore' inherit
from classes of the 'base.ecore' and the test
program checks if the feature id mapping works
for derived features and opposite relations.

The model code and the test code was regenerated
with the changed code generator.

Additional changes:

- A bug was fixed in the Generator code that
  removes all files which are not generated
  during the last run of the generator (option: --clear).
  If a target version is set the start folder
  of the file search must be prepended with
  that target version.