Improve Speed of Emf4cpp Parser
The ReferenceEListImpl::push_back() checks if the
object to add is already contained in the list, and
only appends if not. Creating large lists hence leads
to O(n^2).
Therefore a new method 'push_back_unsafe()' was added
to the EList inheritance hierarchy which omits this
test and can be used as a fast but unsafe alternative.

The check, if an element already exists, was outsourced
to a new method 'EList::contains()'.

The unsafe push_back is now used in the XMLHandler to
improve the parsing speed. It should be used anywhere
to append freshly-created objects to their parent
container.

Additional changes:

The handler and parser that uses the xerces parser was
updated according the handler/parser that uses the
simple parser.

The xerces parser/handler classes were renamed to
XercesHandler/XercesParser to prevent a name collision
with the (simple) parser/handler when build into the
ecorecpp-library. Note, that this is not done actually
to avoid a dependency to the libxerces library.

Change-Id: I2be06ca87140a0d4e7cde3dd497fea6fe950b3f3