Assertion During Reference Serialization
The code, that creates the string representation for a
reference relation both the source and the destination
EObject, needs to live in a EResource, which was tested by
an assertion.

This is now replaced by an exception so that the user can
react on a failing serialization.

A new exception 'ReferenceStringCreationException' was
introduced for that purpose, which derives from
std::logic_error and additionally can be asked
for the source and destination EObject and if the
reference is cross document.

Two try-catch blocks were removed from the XMLSerializer
so that the thrown exceptions find their way to the
caller of the library.

Change-Id: Ic50ec097328ee108204be7a1c579eb33a45a2d2b