Remove Copy Constructor for Class ::ecore::EObject
The explicit copy constructor of class ::ecore::EObject was removed because
it is now obsolete.

The copy constructor was introduced in 6ced2bb3 for a proper implementation
of reference counting for an EObject held by an intrusive_ptr. Since the use
of intrusive_ptr was replaced by a shared_ptr/weak_ptr implementation for a
proper handling of non-containment references this constructor can be removed.

Change-Id: I7460425e8c1a94ac7632f1b79826dc73d53c759c