URIHandler cannot cope with UTF16 encoded strings
When using special characters like german 'umlauts' inside
a file path QUrl the QUrl::toLocalFile() method will return
a QString with utf16 encoding under windows.

The current implementation of the 'create stream' methods of the
URIHandler cannot cope with the encoding of the resulting std::string
when using QString::toStdString().

Therefore, the implementation was hardened by using the new
(c++17) std::filesystem::path for opening the std streams.
These can be initialized by a std::wstring with the proper
encoding.

A compiler version that supports c++17 (gcc 8+) is now mandatory.

Change-Id: Icded0bdfd431488c3f6e5c1b9bc4fbebc855e842